config: download config file. excel: remove "downloading" message
To experiment with system configurations, a way to let the user save and restore configuration from/to file is required. So instead of manually set up every options. So this code provides "download configuration"

Jquery.fileDownload() requires a cookie to be set in order to hide "download is in progress" message box and to  notify download error. So apply this cookie in every needed files ( excel, config, and so ). Notice that fpdf already handles byself this cookie