vwls_Puesto: evaluate puestoFinal instead of puestoParcial
In the second round for a series, it's more important to show final position instead of round position.

So, in order to properly show position regardless which series are we running, we need to re-elaborate "getPuesto() from clasificacionesFunctions", group all rounds in a serie and sum their times and penalizations

TODO: in combo views, recents datagrid may get "undefined" results when redrawing scores datagrid. this should be fixed