getPuesto() prepare all screens... and revert server side code :-(
I've found a missconception error in algorithms for quick evaluate dog ranking: we need to take care on SCT/MCT and -in some circumstances- SCT depends on the current dog... so cannot properly evaluate ranking without "storing" dog results in database

Need to rewrite server code to take care on SCT/MCT and simulate table (ordered) insertion in memory. that is: result is almost the same of retrieve results/classification for _every_ dog

Also, for global ranking, need to do the same process on every round.

So revert changes, but let the code ready to work when algorithm get working