pb(menu)_finales_(equipos) restore reload features
Auto-refresh is disabled in team view when any row is expanded. This is done to improve user experience ( do not close a team being viewed ) cause there is no easy way to store wich teams are expanded, and preserve order in the middle of a contest.

So two ways of re-enable autorefresh in team view:
- collapse all teams
- click on refresh button