videowall & livestream: add round(s) results info for dog being called
Whenever a "llamada" event is processed, in the process to fill current dog data, we need dog results to evaluate required time and penalization to become the first in round.
So need a way to retrieve current dog results. To do so, a little patch is included in getDogByID and windowCall: add results data for the dog in the requested journey