videowall_final: fix final_last_individual-datagrid to follow templates
As final steps for combined videowall  screens, need to fix last-to-be called datagrid either partial/final individual/teams to use new templates. fix scripts and so.

In this step we had to fix how column width is specified, to use percentages and match with easyui convention.
Next comes teams and partial combined datagrid. TODO: verify every screen that uses this template