videowall simplified: fix Result data shown (Position/Elim/Nopr) . dlg_competicion(): fix performance issue
Now  vws reflects changes in Elim/NotPresent matching fields in tablet
Notice that on "Llamada", previows Elim/NoPr status is stored, but not shown, due to "Position" takes precedence in form('load'). This should be fixed

A performance problem was discovered: need to set autorowheight to false in dlg_competicion datagrid in order to show row numbers. As need an special css to fix row height for cell number, it results in multiple rendering call unless autorowheigt set to false