Entrenamientos: finally timespinners work
Had to disable dragndrop when edit mode is on to allow cursor selection on each field.
An error persist: as result is not an mysql timestamp, formater gives an error. so need to smart handle contents on each cell to detect partial declared timestamps and compose result with older data