Chronometer: do not stop in 'llamada' and 'aceptar' unless requested to do. OSD: do not hide competition window if chrono is running
A long standing bug solved: in livestream and videowalls chrono was out of sync with tablet because on 'llamada' or 'accept' chrono is stopped. This is a bug: in some team competitions, chrono should not stop between dogs. Also, if operator clicks "accept" too late videowall and osd looses chrono mark. This is undesirable

So the golden rule: only stop chrono when it is requested to do. Don't try to be smarter.