International: layout country name beside country flag. try 1
Remember that Countries behaves like clubes, so in datagrid, they use to be at right (and irrelevant) side... but for international contests, best place for country name is beside to country flag

So this is a first try to patch datagrid to add a moveField() method to re-organize datagrid columns. works... but datagrid header remains unchanged. So stiil need to work...and wait for help from easyui guys