AuthManager: force nondefault license to enable resetAdminPassword()
Force users of unregistered apps to reinstall, others must provide their license serial number... or call me to get it :-)
This is a way to force people to call me when forget admin password. in this way I can control who really wants to use an specific license. Dirty, but works